Removed all project data from components into projectData folder.
